requirements.txt
Django==2.1.1
django-autoslug-iplweb==1.9.4
django-ckeditor==5.6.1
django-js-asset==1.1.0
pytz==2018.5
